Systems have a setup function that runs after they've been added to the world to setup systems once
Systems have an active property now which can be set and means the system will be skipped. (This way you can disable a system without having to remove it and readd it, but there are times where you should jsut remove a system)

World.add_entity now accepts a list of components to add to that entity simplifying a common patter of adding the entity and then the components

Performance Optimization: When checking for all_components, the system first identifies the component with the smallest set of entities and starts with that set. This significantly reduces the number of comparisons needed, as we only need to check the smallest possible set of entities against other components.